refactor(supabase): update import paths for Database type in client.ts and server.ts
The import paths for the Database type in the client.ts and server.ts files were updated to "../../types/database.types" to match the correct file location. This change improves the accuracy and maintainability of the code.
