+++ utils/supabase/client.ts
@@ -1,4 +1,4 @@
-import { Database } from "@/supabase/database.types";
+import { Database } from "../../types/database.types";
 import { createBrowserClient } from "@supabase/ssr";
 
 export const createClient = () =>
