+++ utils/supabase/server.ts
@@ -1,4 +1,4 @@
-import { Database } from "@/supabase/database.types";
+import { Database } from "../../types/database.types";
 import { createServerClient } from "@supabase/ssr";
 import { cookies } from "next/headers";
